Modularized scripts for quebec
