--- server/util/interfaceTopoJson.ts
@@ -1,33 +1,0 @@
-// Generated by https://quicktype.io
-export interface ITopoJson {
-    type: string;
-    arcs: Array<Array<number[]>>;
-    transform: Transform;
-    objects: Objects;
-}
-
-export interface Objects {
-    circonscriptions: Circonscriptions;
-}
-
-export interface Circonscriptions {
-    type: string;
-    geometries: Geometry[];
-}
-
-export interface Geometry {
-    arcs: Array<Array<number[] | number>>;
-    type: keyof typeof Type;
-    properties: { [key: string]: any }
-}
-
-
-export const Type =  {
-    MultiPolygon: "MultiPolygon",
-    Polygon: "Polygon",
-} as const
-
-export interface Transform {
-    scale: number[];
-    translate: number[];
-}
